feat: add a gitpoll daemon for configs
