--- scripts/config/genotyper/library/lookup_tables/pcr-0.0.1.yaml
@@ -1,7 +1,0 @@
-name: ...
-description: ...
-version: 0.0.1
-
-data:
-  malaria-tlb:
-    ...
