--- scripts/config/genotyper/library/modules-0.0.1.yaml
@@ -1,80 +1,0 @@
-title: ...
-description: ..
-version: 0.0.1
-endpoint: modules/
-
-content:
-  TipRM:
-    info:
-      name: Tip Resource Manager
-      type: Resorce Manager
-      description: Tip RM manages tip boxes with various tip sizes in them
-    api:
-      endpoint: tip_rm/{ID}/
-      variables:
-        ID: uint
-      paths:
-        count/:
-          get:
-            summary: Respond with the counts of tips as well as tip boxes
-            timeout: 1s
-            response:
-              - name: tip_box_capacity
-                type: uint
-                description: ...
-              - name: tip_capacity
-                type: uint
-                description: ...
-              - name: fresh_tip_count
-                type: uint
-                description: ...
-              - name: fresh_tip_box_count
-                type: uint
-                description: ...
-              - name: used_tip_box_count
-                type: uint
-                description: ...
-          book_tip/:
-            post:
-              summary: If tip is available, book a tip for the given sample and do not use use it otherwise
-              timeout: 1s
-              parameter:
-                - name: count
-                  type: uint
-                  description: ...
-              response:
-                - name: confirmation
-                  type: bool
-                  description: ...
-          discard_active_tray/:
-            post:
-              summary: ...
-              timeout: 5m
-              response:
-                - name: confirmation
-                  type: bool
-                  description: ...
-          load_next_tray/:
-            post:
-              summary: ...
-              timeout: 5m
-              response:
-                - name: confirmation
-                  type: bool
-          get_booked_tip/:
-            post:
-              summary: moves the slider of the Tip RM to a location and responds with the location for the gantry to pick up the tip form
-              timeout: 5m
-              response:
-                - name: x
-                  type: float
-                - name: z
-                  type: float
-
-  Gantry:
-    info:
-      name: Gantry Instrument
-      type: Instrument
-      description: Gantry instrument allows motion in XZ axis
-    endpoint: modules/instrument
-
