--- scripts/config/genotyper/library/nodes-0.0.1.yaml
@@ -1,143 +1,0 @@
-title: Nodes
-version: 0.0.1
-description: ...
-
-content:
-  is_tip_available:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: any
-    description: Checks if a tip is is_tip_available
-  
-  tip_stock_error:
-    type: error
-    description: There is an error related to to stock / inventory
-  
-  is_tip_available_in_tray:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: any
-    description: There is an error related to to stock / inventory
-  
-  is_discard_success:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  is_discard_tip_successful:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  tray_maintainance_error:
-    type: error
-    description: There is an error related to to stock / inventory
-  
-  move_slider_to_load_position:
-    type: action
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  load_next_tray:
-    type: action
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  load_new_tray_maintainance_error:
-    type: error
-    description: There is an error related to to stock / inventory
-  
-  is_already_in_position:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  move_tip_slider_to_position:
-    type: action
-    zenoh: 
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  is_slider_position_reached:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  is_load_new_tray_successful:
-    type: condition
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: ...
-  
-  tray_maintainance_error:
-    type: error
-    description: There is an error related to to stock / inventory
-  
-  pickup_tip_using_gantry:
-    type: action
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-  
-  is_caught_tip_firm_and_gantry:
-    type: condition
-    zenoh:
-      modules:
-        - Pipette
-      min_reply: one
-      input_variable: active_pipette
-    description: There is an error related to to stock / inventory
-
-  goto_discard_position:
-    type: action
-    zenoh:
-      modules:
-        - Gantry
-      min_reply: any
-    description: There is an error related to to stock / inventory
-
-  prepare_to_discard:
-    type: action
-    zenoh:
-      modules:
-        - TipRM
-      min_reply: one
-    description: There is an error related to to stock / inventory
-
-  eject_tip:
-    type: action
-    zenoh:
-      modules:
-        - Pipette
-        - TipRM
-      min_reply: any
-    description: ...
-
