--- scripts/config/genotyper/library/tools-0.0.1.yaml
@@ -1,20 +1,0 @@
-version: 0.0.1
-
-content:
-  NoTool:
-    name: No Tool
-    description: Represents no tool mounted on the tool change head
-  
-  Pipette:
-    name: Pipette
-    description: Pipette allows to pick and drop variable volumes of liquid
-    pick_up: NoTool
-  
-    variants:
-      - name: 10uL
-        abbr: 10 
-        preffered_when: The range of liquid to be picked up is between 10 and 100 microliters
-      - name: 100uL
-        abbr: 100
-        preffered_when: The range of liquid to be picked up is between 100 and 1000 microliters
-
