--- scripts/config/genotyper/library/trees/attempt_pickup_tip-0.0.1.yaml
@@ -1,49 +1,0 @@
-title: Get Tip
-version: 0.0.1
-description: This behaviour tree co-ordinates all actions involved in getting a tip
-
-participant_modules:
-  - TipRM
-  - Gantry
-
-tree:
-  name: attempt_pickup_tip
-  children:
-    - name: is_tip_available
-      error: tip_stock_error
-
-    - name: prepare_tip_for_pickup
-      children:
-        - name: is_tip_available_in_tray
-          fallback:
-            - name: discard_current_tray
-              children:
-                - name: is_discard_success
-                  error: tray_maintainance_error
-            - name: load_new_tray
-              children:
-                - name: move_slider_to_load_position
-                - name: load_next_tray
-            - name: is_load_new_tray_successful
-              error: load_new_tray_maintainance_error
-
-        - name: position_tip_slider
-          children:
-            - name: is_already_in_position
-            - name: move_tip_slider_to_position
-            - name: is_slider_position_reached
-              error: tray_maintainance_error
-
-    - name: pick_up_using_gantry
-      children:
-        - name: pickup_tip_using_gantry
-        - name: is_caught_tip_firm_and_oriented
-          fallback:
-            - name: discard_tip
-              children:
-                - name: goto_discard_position
-                - name: prepare_to_discard
-                - name: eject_tip
-                - name: is_discard_tip_successful
-                  error: handle_pickup_failure
-
