--- scripts/config/genotyper/library/workflows/tb-pcr-0.0.1.yaml
@@ -1,26 +1,0 @@
-title: TB PCR
-description: Conduct a PCR based test on a human blood sample for infectious disease for TB
-
-version: 0.0.1
-
-data:
-  - name: annealing_temp
-    description: ...
-    type: float
-  - name: annealing_duration
-    description: ...
-    type: float
-
-workflow:
-  - name: attempt_pickup_tip
-    why: We need a tip in order to pick up the sample and move it to DNA extraction stage
-      #  - name: move_sample_from_storage_to_extractor
-      #    why: We need to pick a sample from the storage in order to process it
-      #  - name: extract_dna
-      #    why: We need DNA from the saliva for this experiment
-      #  - name: amplify_dna
-      #    why: We need to amplify the DNA in our sample, in this workflow this step is also where the detection is completed
-      #  - name: analyse_and_report_result
-      #    why: We need to turn the data generated from the detector to the cloud for further processing and generation of notifications and PDF reports to the end user
-      #  - name: Discard Sample
-      #    why: We do not need this sample anymore, discarding is a good option
