+++ scripts/config_gitpoll/.env.sample
@@ -1,0 +1,1 @@
+GITHUB_ACCESS_TOKEN="ghp_..."
