+++ scripts/config_gitpoll/gitpoll_daemon.py
@@ -1,0 +1,35 @@
+#!/usr/bin/env python
+
+import time
+import logging
+from daemonize import Daemonize
+from gitpoll import gitpoll
+
+PID = "/tmp/gitpoll_daemon.pid"
+
+def init_logger():
+    logger = logging.getLogger(__name__)
+    logger.setLevel(logging.DEBUG)
+    logger.propagate = False
+    fh = logging.FileHandler("/tmp/geno_gitpoll_daemon.log", "w")
+    fh.setLevel(logging.DEBUG)
+    logger.addHandler(fh)
+    formatter = logging.Formatter("%(asctime)s [%(levelname)s]: %(message)s")
+    fh.setFormatter(formatter)
+    keep_fds = [fh.stream.fileno()]
+    return logger, keep_fds
+
+def main():
+    logger.debug("Starting daemon")
+    gs = gitpoll.GitPoll()
+    while(True):
+        gs.poll()
+        logger.debug(gs)
+        time.sleep(60)
+
+if __name__ == "__main__":
+    logger, keep_fds = init_logger()
+    daemon = Daemonize(app="gitpoll_daemon", pid=PID, action=main, keep_fds=keep_fds, logger=logger)
+    print(f"Run `cat {daemon.pid}` to get the pid of this daemon")
+    daemon.start()
+
