+++ scripts/config_gitpoll/requirements.txt
@@ -1,0 +1,2 @@
+gitpython
+Daemonize
