+++ workflow/Cargo.toml
@@ -7,13 +7,19 @@ edition = "2021"
 
 [dependencies]
 async-std = "1.12.0"
+clap = { version = "4.2.4", features = ["derive"] }
 dirs = "5.0.0"
+format_serde_error = "0.3.0"
 futures = "0.3.28"
 log = { version = "0.4.17", features = ["std", "release_max_level_info"]}
 semver = "1.0.17"
 serde = { version = "1.0.159", features = ["derive"] }
 serde_yaml = "0.9.21"
 simplelog = "0.12.1"
+
+[lib]
+name = "workflow"
+path = "src/lib.rs"
 
 [[bin]]
 name = "parse"
