+++ workflow/bin/README.md
@@ -1,22 +1,8 @@
+# Binary Utilities
 
+## Library Loader: `parse.rs`
+```sh
+# Read Help
+cargo run --bin parse -- --help
 ```
-Load Library Nodes
-  |
-  v
-  Load BTrees
-    |
-    v
-    Load Workflows
-      |
-      v
-      intialization complete
-        |
-        v
-        Start accepting samples
-          |
-          v
-       |> Start workflow execution
-       |   |
-       |   v
-       |-- Do this forever
-```
+
